pyproject.toml
[project]
name = "pylak"
version = "0.0.9"
authors = [
  { name="Valter Yde Daugberg", email="valteryde@hotmail.com" },
]
description = "Small library to learn to create 2D games"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]
dependencies = [
  "numpy",
  "pygame",
  "pymunk",
  "pillow"
]

[project.urls]
Homepage = "https://github.com/valteryde/pylak"
Issues = "https://github.com/valteryde/kaxe/pylak"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
